+++ src/main/java/roomescape/controller/web/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.web;
 
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
